fix node imports and remove exposed mightFailFunction
