+++ docs/content/contents.mdx
@@ -67,5 +67,4 @@ course:
           filePath: /jsdocs/functions/Might.mdx
           slug: /functions/Might
           type: documentation
-          
 ---
