+++ examples/makeMightFailBasic.ts
@@ -1,7 +1,7 @@
-import { makeMightFail } from "../src/makeMightFail"
+import { makeMightFail, makeMightFailSync } from "../src/makeMightFail"
 
 const resolve = (value: { message: string }) => Promise.resolve(value)
-async function main() {
+async function asyncMain() {
   const func = makeMightFail(resolve)
   const { error, result } = await func({ message: "success" })
   if (error) {
@@ -11,4 +11,18 @@ async function main() {
   console.log(result.message)
 }
 
-main()
+asyncMain()
+
+
+function syncMain() {
+  const parseJSON = makeMightFailSync(JSON.parse)
+
+  const [error, result] = parseJSON("{invalid}")
+  if (error) {
+    console.error(error)
+    return
+  }
+  console.log(result.message)
+} 
+
+syncMain()
