+++ package.json
@@ -2,18 +2,18 @@
   "name": "might-fail",
   "version": "0.6.1",
   "description": "Return an Either object instead of throwing an exception",
-  "main": "dist/index.js",
+  "main": "dist/cjs/index.js",
+  "module": "dist/index.js",
   "types": "dist/index.d.ts",
+  "type": "module",
   "exports": {
     ".": {
       "import": "./dist/index.js",
-      "require": "./dist/cjs/index.js",
-      "types": "./dist/index.d.ts"
+      "require": "./dist/cjs/index.js"
     },
     "./go": {
       "import": "./dist/go/index.js",
-      "require": "./dist/cjs/go/index.js",
-      "types": "./dist/go/index.d.ts"
+      "require": "./dist/cjs/go/index.js"
     }
   },
   "scripts": {
@@ -55,3 +55,4 @@
     "vitest": "^2.1.1"
   }
 }
+
