+++ src/go/Either.ts
@@ -11,10 +11,10 @@ export type Either<T> =
       result: undefined
     } & [undefined, Error])
   | ({
-      result: T
       error: undefined
+      result: T
     } & [T, undefined])
 
 // Note this array interface is undesireable, we would much rather use just an itterator here but TS doesn't yet support that well enough
-// So it has an array interface and proxies array methods. 
+// So it has an array interface and proxies array methods.
 // See the utils `makeEither` for more information.
