+++ src/go/index.ts
@@ -20,4 +20,12 @@ import { mightFail, mightFailSync, Might, Fail } from "./mightFail"
 import { makeMightFail, makeMightFailSync } from "./makeMightFail"
 
 export { Either, mightFail, makeMightFail, mightFailSync, makeMightFailSync, Might, Fail }
-export default { mightFail, makeMightFail, mightFailSync, makeMightFailSync, Might, Fail }
+const defaultExport = {
+   mightFail,
+   makeMightFail,
+   mightFailSync,
+   makeMightFailSync,
+   Might,
+   Fail
+ }
+ export default defaultExport
