+++ src/index.ts
@@ -1,6 +1,14 @@
 import { type Either } from "./Either"
-import { mightFail, mightFailSync, mightFailFunction, Might, Fail } from "./mightFail"
+import { mightFail, mightFailSync, Might, Fail } from "./mightFail"
 import { makeMightFail, makeMightFailSync } from "./makeMightFail"
 
-export { Either, mightFail, makeMightFail, mightFailSync, makeMightFailSync, mightFailFunction, Might, Fail }
-export default { mightFail, makeMightFail, mightFailSync, makeMightFailSync, mightFailFunction, Might, Fail }
+export { Either, mightFail, makeMightFail, mightFailSync, makeMightFailSync, Might, Fail }
+const defaultExport = {
+  mightFail,
+  makeMightFail,
+  mightFailSync,
+  makeMightFailSync,
+  Might,
+  Fail
+}
+export default defaultExport
