+++ tsconfig.json
@@ -10,6 +10,7 @@
     "rootDir": "./src",
     "strict": true,
     "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
     "moduleResolution": "node"
   },
   "include": [
